[refactor] 닉네임 설정 / 오류 클래스 sealed class로 변경 (#64)
