+++ android/domain/src/main/java/com/teamtripdraw/android/domain/exception/NicknameSetupException.kt
@@ -1,0 +1,5 @@
+package com.teamtripdraw.android.domain.exception
+
+sealed class NicknameSetupException : Exception()
+class InvalidNickNameException(override val message: String?) : NicknameSetupException()
+class DuplicateNickNameException(override val message: String) : NicknameSetupException()
