--- android/domain/src/main/java/com/teamtripdraw/android/domain/exception/nicknameSetup/DuplicateNickNameException.kt
@@ -1,3 +1,0 @@
-package com.teamtripdraw.android.domain.exception.nicknameSetup
-
-class DuplicateNickNameException(override val message: String) : Exception(message)
