--- android/domain/src/main/java/com/teamtripdraw/android/domain/exception/nicknameSetup/InvalidNickNameException.kt
@@ -1,3 +1,0 @@
-package com.teamtripdraw.android.domain.exception.nicknameSetup
-
-class InvalidNickNameException(override val message: String?) : Exception(message)
